[diffusion] Preserve dtype in WanVAE nearest upsample

+++ python/sglang/multimodal_gen/runtime/models/vaes/parallel/wan_common_utils.py
@@ -211,7 +211,7 @@ class WanUpsample(nn.Upsample):
     """
 
     def forward(self, x):
-        return super().forward(x.float()).type_as(x)
+        return super().forward(x)
 
 
 is_first_frame = None
